New command: join <arena> <player>
Added a new command to join an arena.

No parameters: The executing player is added to the queue of a random
arena.

One parameter: The executing player is added to the queue of the given
arena.

Two parameters: The player given by the second parameter is added to
the queue of the given arena.