Storing UUIDs different now
Until now the UUIDs were stored by just putting them into the map. The
MongoDB driver converted them to ByteData. Thus they had to be
converted back if they should be read.

To make it easier to read the database the UUIDs will now be stored as
there string representation.

Also fixed a bug that xp and activeItems of a CannonFighter aren’t
stored if he doesn’t exist.