Changing winner check and adding arena cleanup.
Now every second in the games loop it will be checked whether a player
has won. That cleans up the code that would be used if all related
events were checked.

Also a method to regenerate the world after the game end is
implemented. To do that a new class ModifiedBlock was added to store
the old value of a changed Block. Hopefully that works.