Created Shop system
If you want to create a shop you create a ShopGroup which consists of
Shops for every player wanting to access this shop. That way there can
be personalized versions of the same shop. To create a shop personalized
for a player a ShopMaker has to be given. It will be used to create new
shops, following the rules set in the ShopMakers createShop method.

For a specific shop there is also a ClickHandler needed who handles
clicks on items in the shop. This is realized by calling onItemSelect
and passing an ItemSelectEvent, specially designed for that purpose.