Created CannonConstructor
The CannonConstructor is used to construct a Cannon of a specific type
with the correct levels set from the Database for a player.

If canConstruct returns true the call of construct has to return a valid
Cannon for the map associated with the checked name.

Also Cannons now store their owning player, so that argument isn't
needed anymore on call.