Created Cage class to create cages
This class will be used to create cages at certain locations.

When a new Cage object is created the cage form and its center have to
be specified.

The cage can then be spawned by calling createCage and passing an
ItemStack which has information about material and data of the block to
set.

Non solid blocks will be replaced and stored.

With destroyCage the cage can be removed again. Previously changed
blocks will be restored.

A cage can only be created once at a time. If it’s already spawned and
there is an attempt to create it again an IllegalStateException will be
thrown. The same is happening if you are trying to destroy a not
existing cage.

Each part in the CageForm enum has an array containing Triples with the
offsets where blocks have to be put for that cage form. Currently there
is only PLAYER to create a cage around a player.