Extended CannonConstructor interface
There are 3 new methods which will allow better access to relevant
information:

buyNew takes the CannonFighter who wants to buy a new cannon of that
type and returns a new Cannon created with initial values.

getPrice returns the price to buy that cannon.

getItem returns the ItemStack representing the cannon.