Changed ItemSelectEvent
It is now called ItemInteractEvent.

The constructor now takes an InventoryClickEvent.

There are also two new attributes: InventoryAction action to know what
sort of action happend and rawSlot which holds the raw slot where the
item was.

Also an util method isPickUpAction was added which returns true if the
event action is some sort of pickup.