Added slot amount to CannonFighter
Now there can be a possibility to upgrade the available slots.

The list will be automatically filled to have a size equal to the
available slot amount.

There are lots of new methods:

getActivePositon(name) returns the position of the given cannon in the
enabled list, -1 if it isn’t in it.

getSlots() returns the number of slots the player has

getActiveItem(pos) returns the item in the given slot or null if there
is no one or the index is too huge.

selectCannonToSlot(pos, cannon) puts the cannon with the name into the
given slot. It returns the cannon that was there before or null if
there was no cannon or the wanted position is out of range.