Fixed bug game not quitting correct
This adds two new methods:

GameManager.gameOver(game) notifies the GameManager that the given game
is over and can be cleared. The GameManager will remove the game to
make space for a new one on this arena.

Arena.resetSpawns() this method clears all set spawns and makes them
available again.