Adding new game features and fixing bugs
Now spectators will also see the barriers and receive a message if they
try to leave the arena.

The use of items will be cancelled for players if they are in the game,
cannons are now only usable when the game is ingame.

After game the inventory will be cleared, the health reset, players
will see everybody again and the cannons of the player will be reset
for the next game.