Changed Cannon interface
fire now takes an argument: The item stack that was used. This is
necessary as the item stack is cloned when added to the inventory. With
that step the item can be updated in the inventory.

Also there is a new reset method which will be called when the player
leaves the game. It should reset the cannon to it’s pregame state.