FireballCannon shop now doesn't close on buy
If an upgrade is bought in the FireballCannon shop until now it has
closed. That is very complicated for the user. It will now regenerate
and reopen.

Also cleaned some code. The cannon shop now won’t open the Cannons shop
if it is bought, but instead stay in the Cannon shop. Another click on
the item is needed to open the upgrade shop. This will make it easier
for the player to buy multiple cannons in a row.