Extended CannonFighterDeathEvent
It now additionally has outOfGame which is a boolean showing whether
the player that got killed lost his last life and will be out of game.

That is used to only output the remaining players if a player is out of
game.

Also the reward is now divided by the amount of winners, so every
winner gets just a part.