Added clickable join signs
Now it is possible to create signs which will add you to the queue or
game if you click on them. They are created by putting [CFJoin] on the
second and the arena name on the third row.

Also the worlds spawn location is now set to the lobby location from
the config.