Own server messaging system.
With additional threads there now is a new system for server to
communicate with the proxy. This way no player has to be on the server
to get the message send.

Also the code was cleaned up and structured better now. The game code
now is no longer in the Util part.