Fix various bugs
The comparation of Participants failed because of the missing hashcode.

The IP for requests to the SocketServer will now be checked. If it can
be resolved to localhost that will be done.

Game->Proxy: PlayerLeave now takes the leaving player and the server to
which he will be teleported.

Added outputs

Tried to fix not leaving bug