Major redesign of upgrade storage
As it is very annoying to copy the code for upgrades like cooldown and
ammo into every subclass I tried to make it part of the super class.
The result is a Map consisting of the cannons name as the key and an
UpgradeMap holding the UpgradeLists for the specific Cannon.

An UpgradeMap extends a HashMap<String, UpgradeList<?>> and has the
following additional methods:
setUpgrade(String name, Entry<String, Object> entry, Class<T> type):
Sets the Upgrade serialized in entry with generic type type for the
given upgrade name

setUpgrade(String name, int level, int price, T value, Class<T> type):
The same as the one before except that all values are passed seperately.

setPrice(String name, int level, int price, Class<T> type):
Sets the price of the given level of the given upgrade of type.

setValue(String name, int level, T value, Class<T> type):
Sets the value of the given level of the given upgrade of type.

Upgrade<T> getUpgrade(String name, int level, Class<T> type):
Gets the upgrade for the given level with the given class for the given
upgrade.

int getLevels(String name):
Gets the amount of levels there currently exist for the given upgrade,
0 if it isn’t registered.

Map<String, Object> serialize():
Serializes it to make it storable in the MongoDB.

———————————
These UpgradeMaps consist of UpgradeLists. They extend
ArrayList<Upgrade<T>> and have the following extra functions.
ATTENTION!!!! Only these new methods should be used as only those will
handle the data correctly, list indices start at 0 while the ones for
the UpgradeList in the new methods start at 1, because there is no
level 0…

Upgrade<T> setLevel(int level, Upgrade<T> upgrade):
Sets the Upgrade available at the specific level. If the list hasn't
enough entries they will be filled.
For example if only level 1 and 2 are specified and you try to add
level 5 it will set level 3 and 4 to a clone of what you want to be
level 5.

Upgrade<T> getForLevel(int level):
gets the Upgrade for the given level.

int getLevels():
The amount of existing levels, is list.size().

Map<String, Object> serialize():
Serializes the UpgradeList so it can be stored in a MongoDB

———————————
Upgrade is now clonable. Also it has setters for price and value.

———————————
Use the previously mentioned changes in the abstract Cannon class and
the FireballCannon subclass.

———————————
Hopefully that will work!