ItemInteractEvent now has next shop
A next shop can be set which will be opened after the click if set.

Also the shop no longer closes another shop when opening. This should
result in a not resetting cursor.

Also the closing and opening in the InventoryClickEvent is now made
with one tick delay. This is the way the documentation recommends it, I
don’t know why I did it instantly. It caused strange bugs etc.