Added item stack to UpgradeList
As an upgrade has to be represented by an item in game this is stored
in the UpgradeList. Setter and getter is added.

The upgrade map also has a getter and a setter for this purpose.
UpgradeMap also has a new method getItemMap which will return a Map
with upgrade names as keys and their item representation as value.