Added UpgradeChanger<T> interface
This interface will be used to make it easy to modify upgrades from
inside a chest GUI.

Right now it only has methods to in-/decrease the value fast or slow.

This has to be implemented in a way so the Upgrade of the generic type
can and will be changed.

There are two implementations done already: IntegerUpgradeChanger for
Integers and DoubleUpgradeChanger for Doubles.

Both take two int/double one for slow change, one for fast.