Make shops autounregistering
Now shops will unregister themselves when they are closed. This way a
shop can only be used ones.

I think in addition to that there also were a lot of cases where the
inventory opened wasn’t the one representing this shop. Now this should
be better by checking if the right person is clicking in the inventory
after checking the inventory’s name.

After closing the inventory reusing it will cause an
IllegalStateException.

Also a spelling mistake has been fixed