Upgrade changer now supports the value and initialisaton
Now the interface has a lot new methods. There are ones to de/increase
the price slow/fast and others for the value.

Also there are methods to get the slow/fast price/value change and a
method to get the initial value/price which will be used to create new
upgrades