Created setup for given upgrade level
The shop will be able to modify one given upgrade’s price and value.
Also there are arrows to get to the previous or next upgrade level for
this upgrade. The delete button will delete the level. The level will
be deleted and following levels will get one level down. After that
different things will happen: If you deleted level 1 you will see the
new level 1, which was level 2 before. If you deleted any other level
you will see level - 1, so deleting level 3 will result in seeing level
2, because otherwise level 3 would be readded instantly. Deletion
wouldn’t work as expected… When deleting all levels you will get back
to the Cannon-Setup page. This is the only thing possible right now, as
the UpgradeList doesn’t know for which upgrade name it is used. This
may change in the future.