Adding equals and hashcode to Shop
Also the shop no longer is registered as a listener when it is created,
but instead when the inventory is opened.

The shop stopped working if opening one with the same name or reopening
it, e.g. setup for an upgrade level. I thinks that happens because he
registers one listener for the new shop and then unregistering the one
from the old shop. These both may have the same return of equals. To
avoid unregistering the new shops listener there now is a shopCount
added in the hashCode. A shop now gets an id and increases the shop
count. In equals the shop id’s will be compared.

Hopefully this prevents ambiguation and too early unregistering…