Send disconnect when the server goes offline
Games and Hubs are now using the reconnect system. Both are sending a
shutdown notification to the Proxy, so there shouldn’t be any
EOFExceptions on the Proxyside anymore.

Also a NPE is fixed when disabling the game plugin without a map loaded.

The backup should now be restored correctly.