Fix NPE and stop reconnecting on disable
If no connection could be opened socket is null. Thus closing it would
result in a NPE. This is fixed now by checking if it is actually set.

The games/hubs send a shutdown message to the proxy. After that they
close the socket. This results in a EOFException in the Thread’s run.
There he tries to reconnect to the server but the plugin actually tries
to disable and an IllegalPluginAccessException will be raised. This is
fixed by checking whether the plugin is actually enabled. If not it
won’t try to reconnect.

Fixes #21 and closes #19