Rewrote upgrade storage
Upgrades are now stored in Cannon not by the subclasses. That should
save a lot of duplicate code later on.

This also gives the opportunity to implement the actual upgrade shop in
the Cannon class, also less code.

Some code may seem more complicated but overall it will save tons of
duplicate code later on.