Rewrote arena setup
A arena is now setup with /setup load <name> That will load the
existing arena from the database or create a new one.

The arena border has been removed. It isn’t necessary as the arena will
be the only thing in a world.

Thus the move event doesn’t need to be checked anymore. It also removes
a potential kick because of flying.

Should close #15