New possiblity for new cannons: MovingParticle
The MovingParticle shoots a snowball that gets instantly destroyed,
thus it isn’t visible while the server knows about the snowball and
thus a particle trail can and will be created there.

To handle hits a HitHandler is needed. It will be called when an entity
or a block is hit.