Make MovingParticle more customizable
Now it will call a new method HitHandler#flying(Location) every Tick
which gets the bullets current location passed. This way every cannon
can handle the particles and effects on its own.
This is done in the StandardCannon, it should now have a trail of green
wool block cracks, this is just for testing.