Storage update #1
Now the CannonFighter should be saved less frequently. The new method
save upserts the player into the the database.

Also through deleting the old money system this had to be implemented,
this is done.

The next commits will all be around the internals, won’t add new
features but hopefully make the whole system more reliable.