Make PotionEffets usable
CannonFighter now has addPotionEffect which adds a PotionEffect and a
callback of type PotionEffectOverCallback which has the method
potionEffectEnded. This will be called when the potion effect ends or
gets cancelled and replaced by another effect of the same type. This
way changes made to the player can be shown by setting a potion effect
on hin.