Added IceCannon
This Cannon will slow down the player that gets hit.
Cooldown time, ammo, damage made, radius in which to effect players,
the duration of the effect and the strength of the slow down can be
upgraded.

Added formatPercentage to Language to format a double into percent,
where 1.0 will result in 100 %