Added BlockManager
A BlockManager manages changes made to blocks. It has setBlock to set a
block permanently and setBlockTemporary to set the block for a given
amount of ticks. A CannonFightPlugin now needs a BlockManager, which is
retrieved by getBlockManager.
The Hub will throw an UnsupportedOperationException, as BlockChanging
isn’t supported in the hub.
The Game has a GameBlockManager implementing BlockManager. It handles
block changes by storing a ModifiedBlock with location and type+data in
a BlockChange. A BlockChange is a BukkitRunnable which will just
restore the block by calling restore on the ModifiedBlock.